+++ x-pack/test/api_integration/deployment_agnostic/apis/observability/apm/index.ts
@@ -14,6 +14,7 @@ export default function apmApiIntegrationTests({
     loadTestFile(require.resolve('./agent_explorer'));
     loadTestFile(require.resolve('./custom_dashboards'));
     loadTestFile(require.resolve('./dependencies'));
+    loadTestFile(require.resolve('./entities'));
     loadTestFile(require.resolve('./cold_start'));
   });
 }
